Notebook/requirements.txt
pandas
requests
matplotlib
seaborn
scikit-learn
sentence-transformers
nltk
python-pptx
numpy
scipy